Add files via upload
i have to research the same and share details on a word document with the class.

1. What do we understandfrom this role
2. What are the requirements
3. What are the competencies
4. Which HR alumni works in this company
5. A basic understanding of this company and the industry
6. Any recent updates about this company or the industry
7. Website linkages
8. Industry report linkages.



aim - a small research on company dr lal path labs.
above are the pointers we have to write on the company.
Each student is given a potential recruiter company .
Trainee HR is the job description which i have to write about.. the ctc. Job requirements.